ag_ui_copilotkit_example/backend/requirements.txt
# AG-UI Financial Analysis Server - Python Dependencies

# Core Agent Framework
agent-framework>=1.0.0
agent-framework-ag-ui>=1.0.0
agent-framework-azure>=1.0.0

# Azure SDK
azure-identity>=1.15.0
azure-ai-projects>=1.0.0

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.2
